minor, fix NPE

+++ server-base/src/main/java/org/apache/kylin/rest/controller/ProjectController.java
@@ -193,7 +193,11 @@ public class ProjectController extends BasicController {
         try {
 
             ProjectInstance project = projectService.getProjectManager().getProject(projectName);
-            projectService.deleteProject(projectName, project);
+            if (project != null) {
+                projectService.deleteProject(projectName, project);
+            } else {
+                logger.info("Project {} not exists", projectName);
+            }
         } catch (Exception e) {
             logger.error(e.getLocalizedMessage(), e);
             throw new InternalErrorException("Failed to delete project. " + " Caused by: " + e.getMessage(), e);
